refactor:appBarをクラスに分割
