+++ frontend/where_child_bus_guardian/lib/app.dart
@@ -4,6 +4,7 @@ import 'package:where_child_bus_guardian/pages/check_page/check_page.dart';
 import 'package:where_child_bus_guardian/pages/daily_page/daily_page.dart';
 import 'package:where_child_bus_guardian/pages/map_page/map_page.dart';
 import 'package:where_child_bus_guardian/components/utils/half_circle_painter.dart';
+import 'package:where_child_bus_guardian/components/utils/custom_app_bar.dart';
 import 'package:where_child_bus_guardian/components/utils/custom_bottom_app_bar.dart';
 
 class App extends StatefulWidget {
@@ -19,7 +20,9 @@ class _AppState extends State<App> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: appbar(),
+      appBar: CustomAppBar(
+        selectedIndex: _selectedIndex,
+      ),
       body: Padding(
           padding: EdgeInsets.only(top: 30),
           child: IndexedStack(
@@ -29,45 +32,6 @@ class _AppState extends State<App> {
         selectedIndex: _selectedIndex,
         onItemTapped: (index) => setState(() => _selectedIndex = index),
       ),
-    );
-  }
-
-  AppBar appbar() {
-    return AppBar(
-      backgroundColor: const Color.fromARGB(255, 255, 243, 190),
-      title: Text(''),
-      centerTitle: true,
-      bottom: PreferredSize(
-        preferredSize: Size.fromHeight(0),
-        child: CustomPaint(
-          painter: HalfCirclePainter(),
-          child: Container(
-            height: 0,
-          ),
-        ),
-      ),
-      flexibleSpace: layoutBuilder(),
-    );
-  }
-
-  LayoutBuilder layoutBuilder() {
-    return LayoutBuilder(
-      builder: (BuildContext context, BoxConstraints constraints) {
-        double appBarHeight = constraints.biggest.height;
-        return Center(
-          child: Container(
-            child: Text(
-              ['日々の記録', '地図', '乗車確認'][_selectedIndex],
-              style: const TextStyle(
-                  fontSize: 20.0,
-                  fontWeight: FontWeight.bold,
-                  color: Color.fromARGB(255, 67, 55, 1)),
-            ),
-            // AppBarの高さに応じて位置を調整
-            padding: EdgeInsets.only(top: appBarHeight / 2),
-          ),
-        );
-      },
     );
   }
 }
